🔧 (docker-compose.yaml): move ntopng configuration to configs section
Improved the ntopng service configuration by moving all command line parameters to a dedicated config file. This change makes the configuration more maintainable and easier to read. Added a packet filter to exclude traffic between specific networks (10.0.99.0/24 and 10.0.98.0/24) and changed the input interface from eth1 to a TCP connection on 127.0.0.1:5556. The configuration is now mounted as a file instead of being passed as command line arguments.